Adds native module to start activity and signing wrapper module.
This change set adds a native module that can be used to start an activity for a result. Specifically, this native module will allow us to kick off an intent that will open the app used to sign the completed template. The native module is set up as a promise, so you can await the response from the sign app. It also adds a wrapper to invoke the IXO signing app.